+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added the ability to make pets sit, they wont follow nor defend you in this state
 - Portals that spawn in place of the last staircase at old style dungeons to prevent stair cheesing
 - Mutliple singleplayer worlds and map generation UI.
+- New arena building in desert cities, suitable for PVP, also NPCs like to watch the fights too
 
 ### Changed